Fixed: Code readability (spaces, code comments) and segregate repetitive  tags to individual components
